[refactor/#27] Adapter, Port 구조 Usecase로 리팩터링
